fix: category viewmodel -> video viewmodel 통합
